detect device orientation and remap axes
Detect when phone is in vertical position and remap the axes
accordingly. This way we can continue to receive correct compass
updates using the axis that passes from the phone's rear camera.
If device is in flat position remap the axes again to read from the
axis passing from the top of the device.